+++ source/common/event/dispatcher_impl.cc
@@ -20,7 +20,7 @@
 #include "common/filesystem/watcher_impl.h"
 #include "common/network/connection_impl.h"
 #include "common/network/dns_impl.h"
-#include "common/network/listener_impl.h"
+#include "common/network/tcp_listener_impl.h"
 #include "common/network/udp_listener_impl.h"
 
 #include "event2/event.h"
