+++ source/server/server.cc
@@ -32,9 +32,9 @@
 #include "common/local_info/local_info_impl.h"
 #include "common/memory/stats.h"
 #include "common/network/address_impl.h"
-#include "common/network/listener_impl.h"
 #include "common/network/socket_interface.h"
 #include "common/network/socket_interface_impl.h"
+#include "common/network/tcp_listener_impl.h"
 #include "common/protobuf/utility.h"
 #include "common/router/rds_impl.h"
 #include "common/runtime/runtime_impl.h"
