+++ test/common/network/listener_impl_test.cc
@@ -2,7 +2,7 @@
 #include "envoy/network/exception.h"
 
 #include "common/network/address_impl.h"
-#include "common/network/listener_impl.h"
+#include "common/network/tcp_listener_impl.h"
 #include "common/network/utility.h"
 #include "common/stream_info/stream_info_impl.h"
