+++ test/extensions/filters/listener/proxy_protocol/proxy_protocol_test.cc
@@ -11,8 +11,8 @@
 #include "common/event/dispatcher_impl.h"
 #include "common/network/connection_balancer_impl.h"
 #include "common/network/listen_socket_impl.h"
-#include "common/network/listener_impl.h"
 #include "common/network/raw_buffer_socket.h"
+#include "common/network/tcp_listener_impl.h"
 #include "common/network/utility.h"
 
 #include "server/connection_handler_impl.h"
